django.mk: remove virtualenv.py
This changes the bootstrap script to call `python3 -m venv` instead of
`python virtualenv.py`.

There was one hiccup along the way: travis doesn't include the expected
symlinks in `./ve/bin/`.. I'm not sure why.
https://docs.python.org/3/library/venv.html

This is specific to travis, and should work everywhere else, so I've
added a conditional in here for travis.
https://github.com/travis-ci/travis-ci/issues/9524