django.mk: Remove coverage tool
We don't really use this tool in practice, and I don't really find
automated test coverage reports that useful for now. It only has the
ability to give a very rough view of test coverage. It's better to focus
on just writing good quality tests.

So I'm removing this in the interest of getting rid of an unnecessary
package dependency in our stack.